bugged
some reason my site randomly gets covered with an overlay that doesn't make anything clickable? i thought this was a harmless bug (even if very annoying) but then i mashed click and it brought me to a scam website. might be hijacked by a JS script but i don't know which one